requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
transformers==4.43.3
torch==2.3.1
numpy==1.26.4
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1